Rewrote to use ps2:v2 API and skills checking.
Version 2 of the Planetside 2 API has some significant changes and
improvements. The most significant addition is a functional skills
report that will let us check actual certifications instead of just
checking for the presence of certain items in the characters inventory.

This commit changes the tool to use the ps2:v2 API and skill checks
instead of item checks for all certifications.